Starting a rough design of pytx classes.
Trying to build out a common core of features in a classful way. Built
classes for Malware and Threat Indicator object types.

The generator now returns the appropriate instantiated class instead of
json.

This is still super new and not fully functional. It also takes a very
long time to get any data. For example, if you did the following:

    m = Malware()
    m.setup('<app-id>', '<app-secret'>)
    i = m.objects(text='www.facebook.com')

It will return instantly. Once you try this:

    for x in i:
        print x.id
        break

It takes 15+ seconds to do anything. This is horrible performance. I
haven't looked into it to determine if we are waiting on data (doubtful)
or if our processing is horribly unoptimized.