Fixing issues with .get() and .set().
We are being more loose about .set(). We will allow for a developer to
set whatever attribute they want, but we will only honor the attributes
in _fields when submitting a new object and _changed when submitting
modifications to an object.

Removed the custom __setattr__ from ThreatExchangeMember because it
wasn't useful after this change.

Also fixed .get() which wasn't returning properly for all objects so
things which were valid were being returned as None.