Adding '__raw__' and 'json' to object acquisition.
When querying for objects, you can now pass a dictionary into the
`__raw__` argument. This overrides all other arguments and uses the
provided dictionary as the parameters. This is sort of a "quick and
easy" way of passing in parameters if you already have a dict of them
and know they are accurate (or don't care if they aren't). Example:

d = {'text': 'www.facebook.com'}
foo.objects(__raw__=d)

I also added a 'json' argument to .objects(). This by default is set to
False which means the generator will be used (returning you instantiated
objects for each result). If 'json' is set to True, the generator will
be bypassed and the JSON response will be returned instead. This is
useful if you wish to get everything that was returned and handle it
manually. NOTE: This will *NOT* fetch _next and will assume you are
handling that on your own.