Adjusted arguments for getting data.
The 'json' argument for .objects() is now 'full_response'. If set to
True you will get the full response and will have to deal with parsing
it on your own (including pagination). The default is False, but if set
to True will override the option below.

There is now a new option called 'dict_generator' which is a boolean.
The default is False. When set to False the generator will returned
instantiated objects. When set to True the generator will return a dict
object instead.

Also started expanding the README a bit as documentation is non-existent
for people who want to look into it and test it out.