Verbose server response and added scripts.
Adding in the response text when there's a response other than a 200.

Replicated all of the scripts already written for ThreatExchange but
used pytx instead of urllib. Haven't replicated the post script because
I don't want to post anything yet :)

Added the scripts to setup so that they are installed along with pytx.