Add a request broker.
The broker is an entirely separate class used by the other classes to
interact with the API. It handles the building and sending of the
request as well as the consumption and returning of the response. The
benefit of this is that you can interact with the API without having to
instantiate an object first. The broker can also be used as a
stand-alone class to interact with the API directly without using the
classes.

Update scripts to use the "NWO" of getting objects.