Removing exception from is_timestamp validation.
The `since` and `until` parameters that are available to the API user
are based on PHP's `strtotime` function. Even using python's `dateutil`
library I was unable to accurately reproduce proper results that
`strtotime` was giving for some strings. Without them being equivalent
there's no good way of properly validating if `since` and `until` are
proper values.

Given this I've removed the exception from `is_timestamp` and now it
will always return True, leaving the validation of the value up to the
API consumer when things don't go as they expect. The function is still
there in the event we determine an accurate way of validating these
parameters in the future.