Enforce privacy_types and add properties for malware samples.
Enforce the PRIVACY_TYPE requirement for submitting new Indicators in
pytx and perform a quick check in the UI (which pre-populates the
PRIVACY_TYPE dropdown anyways so unless someone is abusing the JS it
should already work).

Added some helper properties to the Malware object:

    .zfh - return a file handle for the b64 decoded zip file
    .zf  - return the raw b64 decoded zip file
    .rfh - return a file handle for the b64 decoded and unzipped file
    .rf  - return the raw b64 decoded and unzipped file

These properties should be flexible enough for people who want to
utilize the sample without having to do the decode/unzip dance at
whatever stage they are looking to get to.

Fixed a couple other bugs with saving objects to ThreatExchange.