Adding `.new()` and `.send()` methods to graph objects.
The `.new()` classmethod will allow you to send parameters to create a
new graph object. The `.send()` method will allow you to send parameters
to the graph as either a GET or a POST, to either the object URL (ex:
/threat_descriptors/) or the object's details URL (ex: /<id>/). This is
just another flexible way of accessing the graph more than anything.

Given the above, the `.save()` method is now only for instantiated
classes. It requires parameters to send when making the POST request. If
no parameters are provided, we will attempt to determine the changes
made to the class and submit them "as-is" (this may or may not have the
desired effect for the developer, but that is on them).

Bumped to 0.2.0 in versions because this breaks the pytx API.