Add page_limit for a maximum num of page results
"limit - Defines the maximum size of a page of results. The maximum is 1,000/"
https://developers.facebook.com/docs/threat-exchange/reference/apis/threat-descriptors

The limit parameter when making a call directly to the Graph API limits
the number of results on a single page that is returned.

It current limits the total number of objects returned.

I've added a parameter to the `objects` method that allows you to set a
page result limit - `page_limit` - which will be the parameter passed to
`limit`. The original `limit` will continue to function as it had.