Implement retries for requests.
Adds the ability to pass in a `retries` argument to things like
`.objects()`, `.details()`, and `.send()` for use in situations where
you will be paging and need to loop over a large set of data. By default
retries will be None (0) and it will retry for the number you provide if
the page fails to load (not a 200).

This does not solve retries for 200 responses with errors yet. This is
just a first step.