Add logging feature.
With the ability to do retries, it's now more important to allow for
people to log what pytx is doing and thinking.

This adds a `setup_logger` function you can import. You pass the
function a path to a file you want to log to. After that, anything you
do with pytx will log there.

Removed the "TOTAL" feature of the generator. It was causing conflicts
with the `limit` of a request where the two could collide prior to
hitting the appropriate number of results.