Capture error responses better.
When an error occurs, capture the contents of it. We now build up a
dictionary containing all of the error contents from the server response
and we add in the `status_code` and the `url` that the request was made
to. This should make it much easier to parse out and automatically use
within scripts since it's already a dictionary, not a large string that
isn't in JSON.