First step towards adding threat privacy group support.
This adds the ThreatPrivacyGroup class for management of Threat Privacy
Groups. Through inheritance it allows you to add new Threat Privacy
Groups as well as update them. There are helper methods for finding out
which Groups you are a member of, owner of, getting members of a Threat
Privacy Group, as well as set the members of a Threat Privacy Group.