Adjust code to use latest changes for groups and descriptors.
The privacy groups now only have an 'id' so the code now uses that
instead of 'group_id'.

Added support for searching using include_expired, max_confidence, and
min_confidence. Also adjusted the docstrings about owner to note that it
can be comma-delimited to include multiple owners in the search.