Add batch request capability.
This adds the ability to submit batch requests to the Graph.

There is a new `Batch` class which has a `.submit()` classmethod. This
takes an arbitrary number of dictionaries as arguments. Each dictionary
should contain the method, URL (full or relative), and potentially a
body. For kwargs it has detailed in the docstrings that it takes similar
named arguments as methods like `.objects()` insofar as modifying how
the requests library submits your requests (modifying headers, proxy,
verify, retries). It also takes an `include_headers` named argument
which allows you to enable headers in the response.

Added a feature to common class methods like `.objects()` where you can
now pass a `request_dict` argument (boolean). If you do, instead of
actually submitting a request to the Graph, it will instead return a
dictionary which is formatted specifically for passing as an argument to
the `Batch.submit()` method. Example:

    from pytx import Batch, ThreatDescriptor

    print Batch.submit(ThreatDescriptor.objects(text='foo',
                                                request_dict=True),
                       ThreatDescriptor.objects(text='bar',
                                                request_dict=True))

Fixed a bug where "include_expired" was not converting to a string so
when it got encoded by requests it was stripping some of the value.

Bumped copyright year.

Bumped version to 0.4.0 as this is a larger addition to the library.