Support for fetching summary data from connections
We recently added support for fetching ‘summary’ data from the
ThreatIndicator -> ThreatDescriptor connection in the ThreatExchange
RESTful APIs.

The summary object will return rolled up data for the descriptors
attached to a ThreatIndicator.  For now, there’s two summary fields:

  - count (the number of descriptors tied to the indicator)
  - preferred (when there are multiple descriptors, the one that is
‘preferred’)

Why do this?  When you encounter an indicator with multiple
descriptors, you have to do your own munging to handle it gracefully.
Rather than everyone have do the work on their own, the APIs offer a
way to do it for you (if you want it to).

But how does the underlying ThreatExchange API determine the
‘preferred’ one?

Through a set of lightweight, pre-defined heuristics, which can be
selected from in the API request.  Initially, the following heuristics
are supported:

  - MINE_OR_NEWEST: the preferred descriptor should be ‘mine’, if I’ve
submitted one.  failing that, return the newest published descriptor;
  - NEWEST:  return the newest published descriptor;
  - OLDEST:  return the oldest published descriptor;
  - FROM_LIST_OR_NEWEST: the preferred descriptor should be one
published by the member IDs from a provided list (order matters), if
they’ve submitted one.  failing that, return the newest published
descriptor.

We can support more heuristics in time, but started simple from the
methods we’d heard from those rolling their own logic.

There will also be support for summary data on other connections (e.g.
malware to indicators, malware family to variants, etc), so the support
is embedded in the Common connection() code in Pytx.