Adds reference iPython Notebooks
Based on a lot of discussions I’ve had with folks using ThreatExchange,
there’s an interest in tools that make that first sharing of data or an
initial data analysis easier.

This PR adds two ipynb files to perform these common tasks: sharing
data and making sense of the data that you’re able to see.

Happy to share more or build out notebooks that answer other questions
people are looking to solve!