Initial implementation for threat tags.
Threat Tags are independent objects with an 'id' and 'text' field. They
have an edge to 'tagged_objects' so if you include that in your fields
list (which is not there by default when querying the graph) it will
show you the related malware and threat descriptors. Also if you include
'tags' in the fields list when querying for Malware Analyses or Threat
Descriptors, the tags will be available to you.