Adjusted default fields for main objects.
Default fields will now be more sane with providing useful information.
If you do not provide the `fields` argument, we will use the
`_default_fields`...by default...

Updated some documentation bits and the Copyright year.

Added a PopularTags class to vocabulary for people who want to leverage
some of the more popularly-used tags in ThreatExchange.

Fixed a bug introduced in a recent commit which attempted to add a
provided filename while reading from a zip handler. Since the filename
you provide might not be one of the filenames in the zip handler, this
will break as it's meant for reading an existing filename, not writing
to a new filename.

Bumping to 0.5.4.

Closes #158.