[RFC][pytx3] Implement simple normalization + distance on text matching
Summary:
Text matching has been a problem for partners, and we had been
telling them "just use levenshtein distance with 95% match", based on
what would map closest to Facebook's systems. Since then, we realized
there was actually a significant amount of normalization happening first.

We sent partners some rough ideas, and there are plenty of articles on
the internet (and plenty of python libraries) that do some kinds of
normalization and distance.

My goals are the following:
1. The suggested normalization should be simple enough that someone
could easily port it to another language from reading this code.
2. Take on the minimum amount of dependency complexity
3. The highest possible incremental value over raw string comparison.
4. Spend less than 3 hours implementing it

This is attempt at fulfilling those goals.

Test Plan:
I potentially need help coming up with a benchmarking set here
(preferably one I can put in the open source repo or talk about
publicly).

I added some debugging printouts with distance, which are not
in the final version.
```
$ cat ~/samples/raw_text.te
The quick brown fox jumps over the lazy dog, ...
"We the People of the United States, in Order to form a more perfect
 Union, establish Justice, ensure domestic Tranquility, provide for the
 common defence, promote the general Welfare, and secure the Blessings of
 Liberty to ourselves and our Posterity, do ordain and establish this
 Constitution for the United States of
 America.", ...
bball now?, ...

$ pytx3 match text "cdallies5"
DCallies5 - 0 cdallies5
  * >26 thequickbr
  * >259 wethepeopl
  * 6 bballnow

$  pytx3 match text "a bb áll no w"
a bb áll n - 0 abballnow
  * >26 thequickbr
  * >259 wethepeopl
  * 1 bballnow

$ pytx3 match text -T  "We the people of the United States, in order to
form a more perfect union, establish justice, insure domestic
tranquility, provide for the common defense, promote the general
welfare, and secure the blessings of liberty to ourselves and our
posterity, do ordain and establish this Constitution for the United
States of America."
We the peo - 13 wethepeopl
  * >233 thequickbr
  * 2 wethepeopl
  * >260 bballnow
3428971913788104 raw_text media_priority_samples

```