Use faiss index to reconstruct hash in search results
Presently, we need to keep an companion index in memory to convert
the faiss index search results (which are indices in the index) back
to the PDQ hash string.

This change uses the reconstruct method on the two types of faiss indexes
we are using to recompute those hash values on search results. This
should reduce the memory requirements for these classes (and the
eventual size of the classes on disk once we support serializing them)
by around n*32 bytes, where n is the number of hashes in the index.