Support custom ids for PDQ hash indexes
Faiss indexes defaults to assigning sequential ids to the items in the index
based on their add order. But it also allows for using a custom id for each
item. Allowing custom ids will be useful when we use these indexes with PDQ
hashes from threat exchange, where the ThreatExchange ID (an unsigned int64)
can be applied to the PDQ.