Fix hashtool256 output ordering
The word array in the Hash256 type is zeroed on the right boundary,
meaning the zeroth element of the word array represents the last 4
hex characters in the hex. However, both the dumpBits and dumpWords
functions in Hash256 assumed the array was zeroed on the left boundary
of the hash.

This fixes the iteration order for those functions to match the output
of the same tool in Java.