First stab at pull-request, master-push gh actions
For testing, we'll need to merge the PR into master. There are two CI
hooks as part of this PR.
1. PR Validator (runs `terraform fmt` and `terraform validate` without
using a remote backend.
2. Master push validate and test (runs `terraform apply` and creates an
integration test friendly environment with a custom prefix "tftest".

Note: Concurrent execution of workflows (multiple terraform apply) on
the same prefix will have undefined results. This can be solved, if we
see a lot of activity on master, we'll have another stab at this.