Move lambdas to hmalib: a top-level python package
a) Easier to create common pieces like a configured logger
b) More natural to write test cases (debatable). Because you are nudged
     to write more unit-testable pieces.
c) Natural home for hma-cli should that become a reality
d) Will feel unnatural to hardcode credentials or resources

The main build artifacts of the hasher-matcher-actioner directory is the
*code* that we expect partners to run. Potentially deliver this as a single
Docker image with multiple entrypoints.

In that line of thought, the Dockerfile and requirements are *key*. So they
deserve to be at the root and not within the lambdas directory. The lambdas
directory is now nuked.

I plan to have another diff out which will move the terraform files into
their own directory 'iac' or 'terraform'? and clean up the root of the project.

Deployed the new lambda Docker image to a prefixed instance.
```shell
$> make docker
$> make upload_docker
$> terraform apply -var "prefix=$(whoami)"
..yes
```

Subscribed my work email to the topic arn:aws:sns:us-east-1:521978645842:dipanjanm-matches20210223205422088500000002
Uploaded data, verified that index gets created.
Uploaded images, verified that I get an email for matches.