Abandoned attempt to get things running on beanstalk
We got far enough that we created a beanstalk environment, deployed
an application version created from a docker image. However, the
thing didn't start and we could not browse the application over the
internet.

My hunch is to do with security groups. From my previous experiences,
the always-there loading icon which I saw when trying to access the
elb usually points to security groups with strict firewalls.

If we do pick this up, perhaps that's the place to look for.