Root terraform support additional_tags
Summary
---
For accurate cost tracking, we should allow partners to specify tags at one place and pass it along to all the sub-modules we create.

This adds a variable definition and passes along the variable's contents to all submodules known at this time.

Test Plan
---
```
$ terraform apply -var "prefix=$(whoami)"
```

Gave me "Plan: 0 to add, 36 to change, 0 to destroy."
And the output showed tags being added.