script: storm images queue with mock upload events
Summary
---
Because uploading to s3 would induce delays that would adversely affect benchmarks, this script instead mocks uploads to s3 by dropping events on the queue.

Given a queue, an s3_bucket, an s3_prefix and number of messages to drop, the script will get a list of images inside the s3_bucket + prefix combo and then drop messages on the SQS queue in batches of 10.

Test plan
---
```
$ ./storm_images_queue https://queue.amazonaws.com/<acid>/<queue_name> <bucket_name> images/10MB/ 100000
```

was run for real queue names and buckets and I'm attaching screenshots of the lambda runs.