Support image file bytes as input to PDQ hashing
Summary
---
In HMA, we fetch file bytes from s3 and then hash them. Writing them to an interim temp file only to read them again is inefficient and has proven to be a source of errors when the file is not flushed after writes.

This allows directly passing along bytes from the python program to the hasher.

also: Tests and typing were added to existing functions.

Test Plan
---
```
$ py.test
```