[HMA] mypy fixes; black, py.test, mypy in CI; rename dtos -> models (#436)
* tiny: Ignore files my editor needs to work

* DTOs are now models

DTOs are not lightweight anymore. They specify how they are stored and how they will be retrieved. Moving these here as a base for the whole host of models we'll need to create to support actions.

* Typing fixes

Summary
---
Ran mypy along black and realised we've been neglecting types for a while now.

Test Plan
---
```
$ cd hasher-matcher-actioner
$ mypy hmalib
```

Produced errors, but limited to aws_secrets.

* fix typing issues with aws_secrets

* Add lint, type validation, and test for HMA to ci workflow

* [HMA] Github workflows: fix python version and dependency file

* Moore lint fixes and workflow adjustments

* [HMA] Fix test, assert true is true ;)