Add PrivacyGroups to PDQ Match Record
This is information that was getting missed. Unless we have a lookup for all indicator ids, this might get tricky to retain. If an indicator_id (what's currently stored) is present in multiple privacy groups, which privacy groups were matched needs to be retained.

I'm not entirely sure whether the changes in main.tf were required. But without it, I was getting KeyErrors.