Extract primary-bucket out of hashing-data
It is used for so much more, (indexing, fetching etc). It might also be time to reconsider the name for hashing-data module.

This was prompted by a weird dependency cycle in trying to get submission_queue and the bucket refer to each other. This was not strictly necessary, but is a step in the direction.