Bank video members (#805)
* Generalize tab for member addition, support video and photo

* Use fixed aspect ratio for image / video previews

This allows the members page to appear like a grid even if all images are different aspect ratios.

This also fixes the problem where the blur filter would overflow into
the actual area. Screenshot attached.

* Replace blur on hover with explicit eye button clicks

This could have been its own PR, and I tried, but rebase fails, and I
don't have the energy to fix it. If that is important to reviewer, can
be done.

* Use ResponsiveEmbed in previewable dropzone

Use aspect ratio instead of height locking

* Support video previews in BankMembers

* Review feedback

- fix punctuation (remove ! and .) where unnecessary
- Use contentType instead of another type = 'photo' | 'video' type in PreviewableDropzone