Support parameterized counts in stream counters
At this point, only adds support for and test a 'buffer' which receives increments and decrements for both aggregate and parameterized counts. It flushes on-demand (or automatically after every 20 writes).

Next steps would be to identify which parameterized counts we want to maintain and inc them in the lambda's `PipelineTableStreamCounter.update_increments_for_records(...)`

Test Plan
---
Added a unittest with a mock ddb table for counts.

Also verified from a local run of the lambda that counts are incremented.