Run indexer on a schedule
Banks will not emit update events for every member added/removed. Instead indexer will run on a schedule and:
a) if rebuilding, get all datafiles for TE and all members for banks.
b) if updating, get updated datafiles for TE and updated members for banks.

Note, this PR only reads existing s3 datafiles on a schedule. The banks related things will happen as part of a later PR.

Test Plan
---
1. Deployed to my instance.
2. Deleted index/ prefix objects in the s3 bucket
3. Manually ran the indexer
4. Verified that the index was re-created