[banks] PendingBankMemberSignalsIndex -> BankMemberSignalCursorIndex
This far, I was only considering indexer as the consumer of bank member signals. So, the Pending index which was sparse worked out fine. However, there are two complications.

a) rebuilding is not going away,
there needs to be a way to read all bank member signals. All of them even if there are writes between page reads. So SK needs to increment.

b) other consumers (eg. publish bank to threatexchange)
The pending index was meant for a single consumer only and for a one time use. We need something more durable.

The new BankMemberSignalCursorIndex is durable. We don't remove items from it on successfully processing. Since the SK is monotonically increasing, there is no chance of missing out on keys.

Updates to existing signals (remove) will result in bringing the updated item to the end of the index because updated_at == current_time.

You can theoretically use this structure as a feed.