Run indexer on a schedule instead of on events (#872)
* Run indexer on a schedule

Banks will not emit update events for every member added/removed. Instead indexer will run on a schedule and:
a) if rebuilding, get all datafiles for TE and all members for banks.
b) if updating, get updated datafiles for TE and updated members for banks.

Note, this PR only reads existing s3 datafiles on a schedule. The banks related things will happen as part of a later PR.

Test Plan
---
1. Deployed to my instance.
2. Deleted index/ prefix objects in the s3 bucket
3. Manually ran the indexer
4. Verified that the index was re-created

* oops: Missed connecting the event rule and the lambda

Tested that terraform apply now shows the target for the cloudwatch rule.

* oops, oops: Also allow cloudwatch to invoke the lambda

* Use indexing_'trigger'  instead of schedule: consistent with fetcher

* Prevent indexing/fetching triggers from calling all functions

Scope down to the functions they need access to.

* Make indexer cloudwatch_event_rule consistent with fetcher