[hma][ui] Consolidate toasts across the app and control them with useContext
Summary
---
While adding a new error toast, I realized:
-  our toasts were not typed. You only had one type of toast. No failure, success etc.
- also, all pages declared their own containers which made styling impossible and violated DRY

This changeset uses React's useContext to provide interested components an _option_ to simply fire-and-forget notifications. The styling is centrally managed and so is the markup.

This allows you to stack notifications, and centrally control them.

Test Plan
---
tried some actions which trigger toasts, and npm build.

Here are screenshots..