Add an EFS durable file system for hashing lambda
Summary
---
This file system will be used to write content hashes by the hashing lambda so that we can build indexes to use for retroaction.

Test Plan
---
```
$ terraform init
$ terraform apply
```

Added the following snippet to the hashing lambda and deployed.

```python
    with open(f"/mnt/durable-storage/{context.aws_request_id}.file", "w") as f:
        f.write("Hello, World!")

    from os import listdir

    logger.info(listdir("/mnt/durable-storage/"))
```

Hit test a few times, deployed a new version of the lambda and then checked the logs..

Saw the following

```
[INFO]	2022-02-24T03:21:05.008Z	9e057c0d-2027-468a-b797-f261d3964178	['944d7e76-0fb5-4a7b-9104-de86fd90dfc0.file', 'a36b2c1b-1ef4-47d0-aa5c-7f69498c12fd.file', '57e95dc9-efaa-4707-abac-ab1017c7c7ed.file', 'abff2ab6-fd19-47dd-85d8-3bcf2f63c5b7.file', '51191f91-75fc-4f6d-a770-44c46725c054.file', '9e057c0d-2027-468a-b797-f261d3964178.file']
```

See how files accumulate in the file system.