During index building, measure time to fetch data from banks. (#962)
* During index building, measure time to fetch data from banks.

Summary
---
The indexing process has multiple stages. One of the newer ones is getting member signals from banks.

So far, it was not getting measured. Now that we are building 1M+ banks, this becomes a significant part of the indexing duration.

Add a metric name and measure it during index building.

Test Plan
---
Deployed to my personal hma instance.

Attached screenshot of cloudwatch metrics. Notice `get_bank_data` in the graph.

* tiny doc fix recommended by @BarrettOlson