Automate local copy of py-tx; unittest to prevent accidental checkin
Summary
---
While we have had a readme entry and some instructions on how to do this. Automating this is a QoL improvement that I have longed for.

The new script exposes three functions. Swithing to a local copy, switching back to a pypi copy and verifying that we are on the pypi copy.

Test Plan
---
```
$ ./scripts/set_threatexchange_source local
<Dockerfile and setup.py updated. local-threatexchange directory created>

$ ./scripts/set_threatexchange_source pypi
<Dockerfile and setup.py back to normal. local-threatexchange directory deleted>

$ ./scripts/set_threatexchange_source ensure-pypi
<Fails after local, but passes on pypi>
```