SignalExchangeAPI returns patch instead of applying deletes
Version Bump?
---
Should entail a major version bump. A SignalExchangeAPI written without the 'patch' method will not work with the current version.

Summary
---
To make fetch work with a persisted networked storage, we need to stop treating SignalExchangeAPI.naive_fetch_merge as the applicator of fetch_iter() responses.

For the in-memory case, we start calling naive_fetch_merges with nothing in the accumulator and are able to reconstruct the final state.

When a fetch store is backed by network storage or a database, like in banks, the accumulator can't be provided without loading the entire bank into memory.

Instead, we add a method naive_convert_to_signal_type_patch which, instead of applying and absorbing the deletes, returns the deletes as a separate sequence.

Test Plan
---
```
$ py.test
```