Fix divide by zero edge case
If the framesPerSec of a video is <= 0, the program will crash.

The vpdq algorithm is not dependent on the
timestamp and therefore if framesPerSec is 0
then every frame will be hashed. The timestamp
will then equal the frame number of the hash.

This is checked for frameMod, but not for the timestamp.