Use libavcodec directly instead of FFmpeg for CPP
It's better to use libavcodec rather than piping to FFmpeg:

- Avoid process overhead, especially for smaller videos

- Greater control over image and stream parameters

- Avoid vulnerabilities related to CLI

- Remove the need to ask for FFmpeg path which simplifies the CLI and code

- Clear error handling for a failure in any step of hashing

But, unfortunately the Python bindings don't require
libav* to be linked. So, it's necessary to keep the old function
for Python and remove the libav headers. This is done by passing
a define "CYTHON" to the Makefile in setup.py.