Fix Python binding link error and cv2 type error
This fixes issues caused by the POSIX functions being replaced with STL.
pdqhamming was included but never linked.

Also fix a strange bug with cv2 returning a float for width and height.
This is completely unrelated to the
fix above, but it occured during a CI test.
height and width must be int before being passed to hashVideoFileFFMpeg