HMA: Improve debug mode checking
In order to prevent the apscheduler from executing twice in debug mode,
hasher-matcher-actioner checks whether WERKZEUG_RUN_MAIN=="true" before
starting the scheduler, because in debug mode the Flask service will
initialize twice - werkzeug will set that environment variable only for
the "outer" process, meaning that if werkzeug is the WSGI serber this is
a reliable way to only run the scheduler once in debug mode.
However, the "production" version of HMA uses gunicorn as the WSGI
server. Since gunicorn doesn't set that enviornment variable, the
production version of HMA will incorrectly not trigger the scheduler.
This PR fixes this by adding a check for debug mode before checking
for WERKZEUG_RUN_MAIN in the environment.