trim whitespace
this is nicer, because we don't trim ones with sentences. take for
example:
```
This is a sentence[SPACE]
and it breaks back there.
```
We don't ditch this type of formatting.
This would be kiinda hard to do with sed so I ended up trying a python
script which works pretty well